Improve based on Tiago's comment.
