+++ tests/rules/test_mt_force_error.py
@@ -24,32 +24,49 @@ import types
 import unittest
 
 from lsst.ts import salobj, watcher
-from lsst.ts.watcher.rules import MTM2MirrorTemperature
+from lsst.ts.watcher.rules import MTForceError
 from lsst.ts.xml.enums.Watcher import AlarmSeverity
 
 STD_TIMEOUT = 5  # Max time to send/receive a topic (seconds)
 
 
-class MTM2MirrorTemperatureTestCase(unittest.IsolatedAsyncioTestCase):
+class TestData:
+
+    def __init__(
+        self,
+        lut_gravity: list[float],
+        lut_temperature: list[float],
+        applied: list[float],
+        measured: list[float],
+        hardpoint_correction: list[float],
+    ) -> None:
+        self.lutGravity = lut_gravity
+        self.lutTemperature = lut_temperature
+        self.applied = applied
+        self.measured = measured
+        self.hardpointCorrection = hardpoint_correction
+
+
+class MTForceErrorTestCase(unittest.IsolatedAsyncioTestCase):
 
     def setUp(self) -> None:
         salobj.set_random_lsst_dds_partition_prefix()
 
     async def test_constructor(self) -> None:
 
-        schema = watcher.rules.MTM2MirrorTemperature.get_schema()
+        schema = watcher.rules.MTForceError.get_schema()
         assert schema is not None
 
-        rule = MTM2MirrorTemperature(None)
+        rule = MTForceError(None)
         assert len(rule.remote_info_list) == 1
-        assert len(rule.remote_info_list[0].poll_names) == 1
+        assert len(rule.remote_info_list[0].poll_names) == 2
 
     async def test_operation(self) -> None:
         watcher_config_dict = dict(
             disabled_sal_components=[],
             auto_acknowledge_delay=3600,
             auto_unacknowledge_delay=3600,
-            rules=[dict(classname="MTM2MirrorTemperature", configs=[{}])],
+            rules=[dict(classname="MTForceError", configs=[{}])],
             escalation=(),
         )
         watcher_config = types.SimpleNamespace(**watcher_config_dict)
@@ -58,72 +75,53 @@ class MTM2MirrorTemperatureTestCase(unittest.IsolatedAsyncioTestCase):
             domain=controller.domain, config=watcher_config
         ) as model:
 
-            rule_name = "MTM2MirrorTemperature.MTM2"
+            rule_name = "MTForceError.MTM2"
             rule = model.rules[rule_name]
             rule.alarm.init_severity_queue()
 
             await model.enable()
 
-            # 12 ring temperature values and 2 intake temperature values
-            normal_ring = [0.0] * 12
-            normal_intake = [0.0] * 2
-
-            gradient_ring = [0.0] * 12
-            gradient_ring[0] = rule.config.gradient + 1.0
+            # 72 axial actuators and 6 tangent links
+            hardpoint_correction_axial = [1.0] * 72
+            hardpoint_correction_tangent = [1.0] * 6
             test_data_items = [
                 {
-                    "topic": controller.tel_temperature,
+                    "topic": controller.tel_axialForce,
                     "fields": {
-                        "ring": normal_ring,
-                        "intake": normal_intake,
+                        "measured": [0.0] * 72,
+                        "hardpointCorrection": hardpoint_correction_axial,
                     },
                     "expected_severity": AlarmSeverity.NONE,
                 },
                 {
-                    "topic": controller.tel_temperature,
+                    "topic": controller.tel_axialForce,
                     "fields": {
-                        "ring": normal_ring,
-                        "intake": [rule.config.intake + 1.0] * 2,
+                        "measured": [100.0] * 72,
+                        "hardpointCorrection": hardpoint_correction_axial,
                     },
-                    "expected_severity": AlarmSeverity.WARNING,
+                    "expected_severity": AlarmSeverity.SERIOUS,
                 },
                 {
-                    "topic": controller.tel_temperature,
+                    "topic": controller.tel_axialForce,
                     "fields": {
-                        "ring": normal_ring,
-                        "intake": normal_intake,
-                    },
-                    "expected_severity": AlarmSeverity.NONE,
-                },
-                {
-                    "topic": controller.tel_temperature,
-                    "fields": {
-                        "ring": [rule.config.ring + 1.0] * 12,
-                        "intake": normal_intake,
-                    },
-                    "expected_severity": AlarmSeverity.WARNING,
-                },
-                {
-                    "topic": controller.tel_temperature,
-                    "fields": {
-                        "ring": normal_ring,
-                        "intake": normal_intake,
+                        "measured": [0.0] * 72,
+                        "hardpointCorrection": hardpoint_correction_axial,
                     },
                     "expected_severity": AlarmSeverity.NONE,
                 },
                 {
-                    "topic": controller.tel_temperature,
+                    "topic": controller.tel_tangentForce,
                     "fields": {
-                        "ring": gradient_ring,
-                        "intake": normal_intake,
+                        "measured": [100.0] * 6,
+                        "hardpointCorrection": hardpoint_correction_tangent,
                     },
-                    "expected_severity": AlarmSeverity.WARNING,
+                    "expected_severity": AlarmSeverity.SERIOUS,
                 },
                 {
-                    "topic": controller.tel_temperature,
+                    "topic": controller.tel_tangentForce,
                     "fields": {
-                        "ring": normal_ring,
-                        "intake": normal_intake,
+                        "measured": [0.0] * 6,
+                        "hardpointCorrection": hardpoint_correction_tangent,
                     },
                     "expected_severity": AlarmSeverity.NONE,
                 },
@@ -139,3 +137,30 @@ class MTM2MirrorTemperatureTestCase(unittest.IsolatedAsyncioTestCase):
                     assert severity == data_item["expected_severity"]
 
             assert rule.alarm.severity_queue.empty()
+
+    def test_check_out_of_range(self) -> None:
+
+        # Normal
+        data_normal = TestData(
+            [0.0] * 5,
+            [0.0] * 5,
+            [1.0] * 5,
+            [1.0] * 5,
+            [1.0, 0.0, 0.0, 1.0, 1.0],
+        )
+
+        rule = MTForceError(watcher.rules.MTForceError.get_schema())
+        list_actuators = rule._check_out_of_range(data_normal, 1.0)
+        assert list_actuators == []
+
+        # Out of range
+        data_normal = TestData(
+            [0.0] * 5,
+            [0.0] * 5,
+            [1.0] * 5,
+            [10.0] * 5,
+            [1.0, 0.0, 0.0, 1.0, 1.0],
+        )
+
+        list_actuators = rule._check_out_of_range(data_normal, 1.0)
+        assert list_actuators == [0, 3, 4]
